requirements.txt
aiogram==3.15.0
asyncpg==0.30.0
openai==1.54.4
python-dotenv==1.0.1
SQLAlchemy==2.0.36
